+++ packages/inngest/src/koa.ts
@@ -1,3 +1,24 @@
+/**
+ * An adapter for Koa to serve and register any declared functions with Inngest,
+ * making them available to be triggered by events.
+ *
+ * @example
+ * ```ts
+ * const handler = serve({
+ *   client: inngest,
+ *   functions
+ * });
+ *
+ * app.use((ctx) => {
+ *   if (ctx.request.path === "/api/inngest") {
+ *     return handler(ctx);
+ *   }
+ * });
+ * ```
+ *
+ * @module
+ */
+
 import type Koa from "koa";
 import {
   InngestCommHandler,
@@ -5,6 +26,10 @@ import {
 } from "./components/InngestCommHandler";
 import { type SupportedFrameworkName } from "./types";
 
+/**
+ * The name of the framework, used to identify the framework in Inngest
+ * dashboards and during testing.
+ */
 export const frameworkName: SupportedFrameworkName = "koa";
 
 /**
@@ -27,7 +52,12 @@ export const frameworkName: SupportedFrameworkName = "koa";
  *
  * @public
  */
-export const serve = (options: ServeHandlerOptions) => {
+// Has explicit return type to avoid JSR-defined "slow types"
+export const serve = (
+  options: ServeHandlerOptions
+): ((
+  ctx: Koa.ParameterizedContext<Koa.DefaultState, Koa.DefaultContext, unknown>
+) => Promise<void>) => {
   const handler = new InngestCommHandler({
     frameworkName,
     ...options,