+++ packages/inngest/src/nuxt.ts
@@ -1,3 +1,10 @@
+/**
+ * An adapter for Nuxt to serve and register any declared functions with
+ * Inngest, making them available to be triggered by events.
+ *
+ * @module
+ */
+
 import {
   type InternalServeHandlerOptions,
   type ServeHandlerOptions,
@@ -5,6 +12,10 @@ import {
 import { serve as serveH3 } from "./h3";
 import { type SupportedFrameworkName } from "./types";
 
+/**
+ * The name of the framework, used to identify the framework in Inngest
+ * dashboards and during testing.
+ */
 export const frameworkName: SupportedFrameworkName = "nuxt";
 
 /**
@@ -13,7 +24,10 @@ export const frameworkName: SupportedFrameworkName = "nuxt";
  *
  * @public
  */
-export const serve = (options: ServeHandlerOptions) => {
+// Has explicit return type to avoid JSR-defined "slow types"
+export const serve = (
+  options: ServeHandlerOptions
+): ReturnType<typeof serveH3> => {
   const optsOverrides: InternalServeHandlerOptions = {
     ...options,
     frameworkName,