+++ packages/inngest/test/benchmark/execution/index.ts
@@ -1,0 +1,11 @@
+import { group } from "mitata";
+import { loadBenchmarks } from "../util";
+
+// Give me Bun
+export const register = async () => {
+  const benchmarks = await loadBenchmarks(["memoized-steps"]);
+
+  group("execution", () => {
+    benchmarks.register();
+  });
+};